fix: use `u64` for parsing Linux partition blocks (#3873)

+++ yazi-fs/src/mounts/linux.rs
@@ -66,7 +66,7 @@ impl Partitions {
 			let mut it = line.split_whitespace();
 			let Some(Ok(_major)) = it.next().map(|s| s.parse::<u16>()) else { continue };
 			let Some(Ok(_minor)) = it.next().map(|s| s.parse::<u16>()) else { continue };
-			let Some(Ok(_blocks)) = it.next().map(|s| s.parse::<u32>()) else { continue };
+			let Some(Ok(_blocks)) = it.next().map(|s| s.parse::<u64>()) else { continue };
 			if let Some(name) = it.next() {
 				set.insert(Self::unmangle_octal(name).into_owned());
 			}
